Update confusion_generator.py
Added the possibility of multiple words to add confusion words.

Main idea is that at some point students might learn to see that there is always two forms of a specific word, and this method allows more flexibility. To note, it's not optimized, and it simply calls the original method twice, which can result in picking the same word twice, but works about 70ms for 2 words and 100 ms for 3,  so it seems like it would be OK as long as not too many words are selected.